Move addReactionToMessage to slack.ts
